+++ src/equations/equations.jl
@@ -64,6 +64,29 @@ direction `orientation` for the coressponding set of governing `equations`
 function flux end
 
 
+"""
+    rotate_to_x(u, normal, equations)
+
+Apply the rotation that maps `normal` onto the x-axis to the convservative variables `u`.
+This is used by [`FluxRotated`](@ref) to calculate the numerical flux of rotationally
+invariant equations in arbitrary normal directions.
+
+See also: [`rotate_from_x`](@ref)
+"""
+function rotate_to_x end
+
+"""
+    rotate_from_x(u, normal, equations)
+
+Apply the rotation that maps the x-axis onto `normal` to the convservative variables `u`.
+This is used by [`FluxRotated`](@ref) to calculate the numerical flux of rotationally
+invariant equations in arbitrary normal directions.
+
+See also: [`rotate_to_x`](@ref)
+"""
+function rotate_from_x end
+
+
 # set sensible default values that may be overwritten by specific equations
 have_nonconservative_terms(::AbstractEquations) = Val(false)
 have_constant_speed(::AbstractEquations) = Val(false)