+++ src/equations/lattice_boltzmann_3d.jl
@@ -226,7 +226,7 @@ end
 
 
 # Calculate 1D flux in for a single point
-@inline function flux(u, orientation, equations::LatticeBoltzmannEquations3D)
+@inline function flux(u, orientation::Integer, equations::LatticeBoltzmannEquations3D)
   if orientation == 1 # x-direction
     v_alpha = equations.v_alpha1
   elseif orientation == 2 # y-direction