+++ src/equations/linear_scalar_advection_1d.jl
@@ -173,7 +173,7 @@ end
 
 
 # Calculate 1D flux in for a single point
-@inline function flux(u, orientation, equation::LinearScalarAdvectionEquation1D)
+@inline function flux(u, orientation::Integer, equation::LinearScalarAdvectionEquation1D)
   a = equation.advectionvelocity[orientation]
   return a * u
 end