+++ test/test_examples_2d_part1.jl
@@ -25,8 +25,8 @@ isdir(outdir) && rm(outdir, recursive=true)
   # Compressible Euler
   include("test_examples_2d_euler.jl")
 
-  # Structured mesh
-  include("test_examples_2d_structured.jl")
+  # Curved mesh
+  include("test_examples_2d_curved.jl")
 end
 
 # Coverage test for all initial conditions